render_sim/render_genesis/requirements.txt
absl-py==2.3.1
addict==2.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
coacd==1.0.7
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.2.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
etils==1.5.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.13.1
Flask==3.1.1
fonttools==4.59.0
freetype-py==2.5.1
fsspec==2024.6.1
genesis-world==0.2.1
glfw==2.9.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lazy_loader==0.4
libigl==2.4.1
llvmlite==0.43.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
mujoco==3.2.5
mypy_extensions==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
open3d==0.19.0
opencv-python==4.11.0.86
OpenEXR==3.3.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
plotly==5.24.1
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.9.2
pydantic==2.7.1
pydantic_core==2.18.2
PyGEL3D==0.6.1
pyglet==2.1.6
pygltflib==1.16.0
Pygments==2.19.2
pymeshlab==2023.12.post3
PyOpenGL==3.1.9
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvista==0.46.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rich==14.1.0
rpds-py==0.27.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
scooby==0.10.1
screeninfo==0.8.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
taichi==1.7.3
tenacity==9.1.2
tetgen==0.6.4
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vtk==9.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
zipp==3.23.0
